Joining our  songsHTML array so that it can be readable and make sense
Chain the join() method to your map() method and pass in an empty string for the separator.


The join() method is used to concatenate all the elements of an array into a single string. It takes an optional parameter called a separator which is used to separate each element of the array. For example: