Setting location property for the audio.
Inside the playSong function, we set the audio.src property equal to song.src. This tells the audio element where to find the audio data for the selected song.

Also, we set the audio.title property equal to song.title. This tells the audio element what to display as the title of the song.